requirements.txt
fastapi>=0.116.0
uvicorn>=0.27.0
psycopg2>=2.9.10
python-dotenv>=1.0.0
mcp[cli]>=1.10.1
async-lru>=2.0.5
asyncpg>=0.30.0 